requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.3
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.40
urllib3==1.26.12
Werkzeug==2.2.2
